# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mentis Semear API",
        "description": "API do monólito Mentis Semear (autenticação JWT, workshops, check-ins, avaliações e diário).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Mesma origem (dev e produção)"
        }
    ],
    "paths": {
...
        "/api/check-ins": {
            "get": {
                "tags": [
                    "CheckIns"
                ],
                "summary": "Lista check-ins (paginado)",
                "operationId": "4ec1707208d5361e97dd9e6fb8142140",
                "parameters": [
                    {
                        "name": "workshop_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CheckIn"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso restrito ao perfil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "CheckIns"
                ],
                "summary": "Registra um check-in",
                "operationId": "73cd5eb6388344a74bfa04936b51691a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "workshop_id",
                                    "name",
                                    "position",
                                    "sector",
                                    "lgpd_read",
                                    "cpf",
                                    "birthday",
                                    "gender",
                                    "celphone"
                                ],
                                "properties": {
                                    "workshop_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "João Participante"
                                    },
                                    "position": {
                                        "type": "string",
                                        "example": "Analista"
                                    },
                                    "sector": {
                                        "type": "string",
                                        "example": "TI"
                                    },
                                    "lgpd_read": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "12345678901"
                                    },
                                    "birthday": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-05-20"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "example": "M"
                                    },
                                    "celphone": {
                                        "type": "string",
                                        "example": "11999999999"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CheckIn"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso restrito ao perfil"
                    },
                    "422": {
                        "description": "Falha de validação (CPF já registrado neste workshop, etc.)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/public/check-ins": {
            "post": {
                "tags": [
                    "Público"
                ],
                "summary": "Check-in público de participante (sem autenticação)",
                "operationId": "9d346e66444a6f3c5f0518e57373bdb1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "workshop_id",
                                    "name",
                                    "position",
                                    "sector",
                                    "lgpd_read",
                                    "cpf",
                                    "birthday",
                                    "gender",
                                    "celphone"
                                ],
                                "properties": {
                                    "workshop_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "João Participante"
                                    },
                                    "position": {
                                        "type": "string",
                                        "example": "Analista"
                                    },
                                    "sector": {
                                        "type": "string",
                                        "example": "TI"
                                    },
                                    "lgpd_read": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "12345678901"
                                    },
                                    "birthday": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-05-20"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "example": "M"
                                    },
                                    "celphone": {
                                        "type": "string",
                                        "example": "11999999999"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Check-in registrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CheckIn"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Falha de validação (CPF já registrado neste workshop, etc.)"
                    }
                }
            }
        },
        "/api/check-ins/{checkIn}": {
            "get": {
                "tags": [
                    "CheckIns"
                ],
                "summary": "Detalha um check-in",
                "operationId": "3f89c68bde7f3dced3779cd0391d6a4b",
                "parameters": [
                    {
                        "name": "checkIn",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Check-in",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CheckIn"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso restrito ao perfil"
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "CheckIns"
                ],
                "summary": "Atualiza um check-in (campos parciais)",
                "operationId": "24dbbefc3c1813039c8197cd16d675bd",
                "parameters": [
                    {
                        "name": "checkIn",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "position": {
                                        "type": "string"
                                    },
                                    "sector": {
                                        "type": "string"
                                    },
                                    "lgpd_read": {
                                        "type": "boolean"
                                    },
                                    "cpf": {
                                        "type": "string"
                                    },
                                    "birthday": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "gender": {
                                        "type": "string"
                                    },
                                    "celphone": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CheckIn"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso restrito ao perfil"
                    },
                    "404": {
                        "description": "Não encontrado"
                    },
                    "422": {
                        "description": "Falha de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "CheckIns"
                ],
                "summary": "Remove um check-in",
                "operationId": "dce9f0f69f2be007340b748f4a472371",
                "parameters": [
                    {
                        "name": "checkIn",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Removido"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso restrito ao perfil"
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CheckIn": {
                "properties": {
                    "id": {
                        "description": "OpenAPI schema only — never instantiated.",
                        "type": "integer",
                        "example": 1
                    },
                    "workshop_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "João Participante"
                    },
                    "position": {
                        "type": "string",
                        "example": "Analista"
                    },
                    "sector": {
                        "type": "string",
                        "example": "TI"
                    },
                    "lgpd_read": {
                        "type": "boolean",
                        "example": true
                    },
                    "lgpd_consent_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "cpf": {
                        "type": "string",
                        "example": "12345678901"
                    },
                    "birthday": {
                        "type": "string",
                        "format": "date",
                        "example": "1990-05-20"
                    },
                    "gender": {
                        "type": "string",
                        "example": "M"
                    },
                    "celphone": {
                        "type": "string",
                        "example": "11999999999"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Envie o token no header: Authorization: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}